init shift dvrp schedules explicitly before initialization of vehicle ques in the shift dispatcher to avoid wrong order of initialization
