+++ contribs/drt/src/main/java/org/matsim/contrib/drt/extension/shifts/dispatcher/DrtShiftDispatcherImpl.java
@@ -84,6 +84,7 @@ public class DrtShiftDispatcherImpl implements DrtShiftDispatcher {
         this.eventsManager = eventsManager;
         this.configGroup = configGroup;
 
+		shiftTaskScheduler.initSchedules();
         createQueues();
     }
